@@ -19,28 +19,28 @@
   },
   "dependencies": {
     "@popperjs/core": "^2.11.8",
-    "ace-builds": "^1.32.0",
+    "ace-builds": "^1.32.5",
     "bootstrap": "^5.3.2",
-    "css-loader": "^6.8.1",
+    "css-loader": "^6.10.0",
     "file-loader": "^6.2.0",
     "jquery": "^3.7.1",
     "jquery-ui": "^1.13.2",
     "jquery-ui-css": "^1.11.5",
     "js-yaml": "^4.1.0",
     "monaco-editor": "0.45.0",
-    "style-loader": "^3.3.3",
-    "tinymce": "^6.8.1"
+    "style-loader": "^3.3.4",
+    "tinymce": "^6.8.2"
   },
   "devDependencies": {
     "@types/bootstrap": "^5.2.10",
-    "@types/jquery": "^3",
+    "@types/jquery": "^3.5.29",
     "@types/rimraf": "^4.0.5",
     "build-if-changed": "^1.5.5",
     "grunt": "^1.6.1",
     "monaco-editor-webpack-plugin": "^7.1.0",
     "npm-run-all": "^4.1.5",
     "rimraf": "^5.0.5",
-    "webpack": "^5.89.0",
+    "webpack": "^5.90.0",
     "webpack-cli": "^5.1.4"
   }
 }
