@@ -1,7 +1,7 @@
 {
   "name": "@craftercms/uppy",
   "description": "CrafterCMS UI Uppy extensions.",
-  "version": "4.1.2",
+  "version": "4.2.0",
   "license": "MIT",
   "main": "lib/index.js",
   "types": "types/index.d.ts",
@@ -34,12 +34,12 @@
     "watch": "babel src -d lib --watch"
   },
   "devDependencies": {
-    "@babel/cli": "^7.23.4",
+    "@babel/cli": "^7.23.9",
     "@babel/plugin-transform-object-assign": "^7.23.3",
-    "@types/eslint": "^8.44.8",
+    "@types/eslint": "^8.56.2",
     "babel-plugin-inline-package-json": "^2.0.0",
     "build-if-changed": "^1.5.5",
-    "eslint": "^8.55.0"
+    "eslint": "^8.56.0"
   },
   "peerDependencies": {
     "uppy": "^2.13.8"
