@@ -19,22 +19,27 @@ import StandardAction from '../../models/StandardAction';
 import { Dispatch } from 'redux';
 import { useDispatch } from 'react-redux';
 import { isPlainObject } from '../../utils/object';
-import { useSnackbar } from 'notistack';
+import { SnackbarKey, useSnackbar } from 'notistack';
 import { getHostToHostBus } from '../../utils/subjects';
-import { newProjectReady, showSystemNotification } from '../../state/actions/system';
+import { blockUI, newProjectReady, showSystemNotification, unblockUI } from '../../state/actions/system';
 import Launcher from '../Launcher/Launcher';
 import useSelection from '../../hooks/useSelection';
 import { useWithPendingChangesCloseRequest } from '../../hooks/useWithPendingChangesCloseRequest';
 import MinimizedBar from '../MinimizedBar';
 import { RenameAssetDialog } from '../RenameAssetDialog';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import Button from '@mui/material/Button';
 import { getSystemLink } from '../../utils/system';
 import useEnv from '../../hooks/useEnv';
 import { filter, map, switchMap } from 'rxjs/operators';
 import { ProjectLifecycleEvent } from '../../models/ProjectLifecycleEvent';
 import { fetchAll as fetchSitesService } from '../../services/sites';
+import IconButton from '@mui/material/IconButton';
+import CloseRounded from '@mui/icons-material/CloseRounded';
+import useAuth from '../../hooks/useAuth';
+import useActiveSiteId from '../../hooks/useActiveSiteId';
 
+// region const ... = lazy(() => import('...'));
 const ViewVersionDialog = lazy(() => import('../ViewVersionDialog'));
 const CompareVersionsDialog = lazy(() => import('../CompareVersionsDialog'));
 const RejectDialog = lazy(() => import('../RejectDialog'));
@@ -64,6 +69,7 @@ const PathSelectionDialog = lazy(() => import('../PathSelectionDialog'));
 const UnlockPublisherDialog = lazy(() => import('../UnlockPublisherDialog'));
 const WidgetDialog = lazy(() => import('../WidgetDialog'));
 const CodeEditorDialog = lazy(() => import('../CodeEditorDialog'));
+// endregion
 
 // @formatter:off
 function createCallback(action: StandardAction, dispatch: Dispatch): (output?: unknown) => void {
@@ -106,9 +112,12 @@ function GlobalDialogManager() {
   const state = useSelection((state) => state.dialogs);
   const contentTypesBranch = useSelection((state) => state.contentTypes);
   const versionsBranch = useSelection((state) => state.versions);
-  const { enqueueSnackbar } = useSnackbar();
+  const { enqueueSnackbar, closeSnackbar } = useSnackbar();
   const dispatch = useDispatch();
-  const { authoringBase } = useEnv();
+  const { authoringBase, socketConnected } = useEnv();
+  const { active: authActive } = useAuth();
+  const activeSiteId = useActiveSiteId();
+  const { formatMessage } = useIntl();
 
   useEffect(() => {
     const hostToHost$ = getHostToHostBus();
@@ -164,6 +173,68 @@ function GlobalDialogManager() {
       });
     return () => subscription.unsubscribe();
   }, [authoringBase, enqueueSnackbar]);
+
+  useEffect(() => {
+    const isIframe = window.location !== window.parent.location;
+    if (!isIframe && authActive && !socketConnected && activeSiteId !== null) {
+      let timeout: NodeJS.Timeout, key: SnackbarKey;
+      timeout = setTimeout(() => {
+        fetch(`${authoringBase}/help/socket-connection-error`)
+          .then((r) => r.text())
+          .then(() => {
+            key = enqueueSnackbar(<FormattedMessage defaultMessage="Studio will continue to retry the connection." />, {
+              variant: 'warning',
+              persist: true,
+              anchorOrigin: { vertical: 'bottom', horizontal: 'center' },
+              alertTitle: <FormattedMessage defaultMessage="Connection with the server interrupted" />,
+              action: (key) => (
+                <>
+                  <Button
+                    href={`${authoringBase}/help/socket-connection-error`}
+                    target="_blank"
+                    size="small"
+                    color="inherit"
+                  >
+                    <FormattedMessage defaultMessage="Learn more" />
+                  </Button>
+                  <IconButton size="small" color="inherit" onClick={() => closeSnackbar(key)}>
+                    <CloseRounded />
+                  </IconButton>
+                </>
+              )
+            });
+          })
+          .catch(() => {
+            dispatch(
+              blockUI({
+                title: formatMessage({ defaultMessage: 'Connection with the server interrupted' }),
+                message: formatMessage({
+                  defaultMessage:
+                    'Studio servers might be down, being restarted or your network connection dropped. Check your connection or ask the administrator to validate server status.'
+                })
+              })
+            );
+          });
+      }, 5000);
+      return () => {
+        clearTimeout(timeout);
+        if (key) {
+          closeSnackbar(key);
+        } else {
+          dispatch(unblockUI());
+        }
+      };
+    }
+  }, [
+    authoringBase,
+    authActive,
+    closeSnackbar,
+    enqueueSnackbar,
+    socketConnected,
+    dispatch,
+    formatMessage,
+    activeSiteId
+  ]);
 
   return (
     <Suspense fallback="">
