@@ -44,6 +44,10 @@ export const pathNavigatorTreeBackgroundRefresh = /*#__PURE__*/ createAction<{ i
   'PATH_NAV_TREE_BACKGROUND_REFRESH'
 );
 
+export const pathNavigatorTreeBulkRefresh = /*#__PURE__*/ createAction<{
+  requests: PayloadWithId<{ backgroundRefresh?: boolean }>[];
+}>('PATH_NAV_TREE_BULK_REFRESH');
+
 export const pathNavigatorTreeRestore = /*#__PURE__*/ createAction<PayloadWithId<{}>>('PATH_NAV_TREE_RESTORE');
 
 export type PathNavigatorTreeRestoreCompletePayload = PayloadWithId<{
@@ -57,8 +61,17 @@ export const pathNavigatorTreeRestoreComplete = /*#__PURE__*/ createAction<PathN
   'PATH_NAV_TREE_RESTORE_COMPLETE'
 );
 
+export const pathNavigatorTreeBulkRestoreComplete = /*#__PURE__*/ createAction<{
+  trees: PathNavigatorTreeRestoreCompletePayload[];
+}>('PATH_NAV_TREE_BULK_RESTORE_COMPLETE');
+
 export const pathNavigatorTreeRestoreFailed = /*#__PURE__*/ createAction<{
   id: string;
+  error: Omit<AjaxError, 'request' | 'xhr'>;
+}>('PATH_NAV_TREE_RESTORE_FAILED');
+
+export const pathNavigatorTreeBulkRestoreFailed = /*#__PURE__*/ createAction<{
+  ids: string[];
   error: Omit<AjaxError, 'request' | 'xhr'>;
 }>('PATH_NAV_TREE_RESTORE_FAILED');
 
@@ -88,18 +101,43 @@ export const pathNavigatorTreeFetchPathPageFailed = /*#__PURE__*/ createAction<{
   error: Omit<AjaxError, 'request' | 'xhr'>;
 }>('PATH_NAV_TREE_FETCH_PATH_PAGE_FAILED');
 
-export const pathNavigatorTreeFetchPathChildren = /*#__PURE__*/ createAction<
-  PayloadWithId<{ path: string; options?: Partial<GetChildrenOptions>; expand?: boolean }>
->('PATH_NAV_TREE_FETCH_PATH_CHILDREN');
+export type PathNavTreeFetchPathChildrenPayload = PayloadWithId<{
+  path: string;
+  options?: Partial<GetChildrenOptions>;
+  expand?: boolean;
+}>;
+
+export const pathNavigatorTreeFetchPathChildren = /*#__PURE__*/ createAction<PathNavTreeFetchPathChildrenPayload>(
+  'PATH_NAV_TREE_FETCH_PATH_CHILDREN'
+);
+
+export type PathNavTreeBulkFetchPathChildrenPayload = { requests: Array<PathNavTreeFetchPathChildrenPayload> };
+
+export const pathNavigatorTreeBulkFetchPathChildren =
+  /*#__PURE__*/ createAction<PathNavTreeBulkFetchPathChildrenPayload>('PATH_NAV_TREE_BULK_FETCH_PATH_CHILDREN');
 
 export const pathNavigatorTreeFetchPathChildrenComplete = /*#__PURE__*/ createAction<
   PayloadWithId<{ children: GetChildrenResponse; parentPath: string; options?: Partial<GetChildrenOptions> }>
 >('PATH_NAV_TREE_FETCH_PATH_CHILDREN_COMPLETE');
 
+export type PathNavTreeBulkFetchPathChildrenCompletePayload = {
+  paths: PayloadWithId<{ children: GetChildrenResponse; parentPath: string; options?: Partial<GetChildrenOptions> }>[];
+};
+
+export const pathNavigatorTreeBulkFetchPathChildrenComplete =
+  /*#__PURE__*/ createAction<PathNavTreeBulkFetchPathChildrenCompletePayload>(
+    'PATH_NAV_BULK_TREE_FETCH_PATH_CHILDREN_COMPLETE'
+  );
+
 export const pathNavigatorTreeFetchPathChildrenFailed = /*#__PURE__*/ createAction<{
   id: string;
   error: Omit<AjaxError, 'request' | 'xhr'>;
 }>('PATH_NAV_TREE_FETCH_PATH_CHILDREN_FAILED');
+
+export const pathNavigatorTreeBulkFetchPathChildrenFailed = /*#__PURE__*/ createAction<{
+  ids: string[];
+  error: Omit<AjaxError, 'request' | 'xhr'>;
+}>('PATH_NAV_TREE_BULK_FETCH_PATH_CHILDREN_FAILED');
 
 export const pathNavigatorTreeUpdate = /*#__PURE__*/ createAction<
   PayloadWithId<{
