@@ -39,6 +39,7 @@ import QuickCreateItem from '../../models/content/QuickCreateItem';
 import StandardAction from '../../models/StandardAction';
 import { AjaxError } from 'rxjs/ajax';
 import {
+  pathNavigatorBulkFetchPathComplete,
   pathNavigatorConditionallySetPathComplete,
   pathNavigatorFetchParentItemsComplete,
   pathNavigatorFetchPathComplete
@@ -48,6 +49,8 @@ import { createLookupTable, reversePluckProps } from '../../utils/object';
 import { SandboxItem } from '../../models/Item';
 import { changeSiteComplete } from '../actions/sites';
 import {
+  pathNavigatorTreeBulkFetchPathChildrenComplete,
+  pathNavigatorTreeBulkRestoreComplete,
   pathNavigatorTreeFetchPathChildrenComplete,
   pathNavigatorTreeFetchPathPageComplete,
   pathNavigatorTreeRestoreComplete,
@@ -91,6 +94,20 @@ const updateItemByPath = (state: ContentState, { payload }) => {
   };
 };
 
+const updateItemsByPaths = (state: ContentState, { payload: { paths } }) => {
+  let nextByPath = state.itemsByPath;
+  paths.forEach((path) => {
+    nextByPath = {
+      ...nextByPath,
+      ...updateItemByPath({ ...state, itemsByPath: nextByPath }, { payload: path }).itemsByPath
+    };
+  });
+  return {
+    ...state,
+    itemsByPath: nextByPath
+  };
+};
+
 const updateItemsBeingFetchedByPath = (state: ContentState, { payload: { path } }) => {
   state.itemsBeingFetchedByPath[path] = true;
 };
@@ -99,6 +116,27 @@ const updateItemsBeingFetchedByPaths = (state, { payload: { paths } }) => {
   paths.forEach((path) => {
     state.itemsBeingFetchedByPath[path] = true;
   });
+};
+
+const updateItemsFromRestoredTree = (state, payload: PathNavigatorTreeRestoreCompletePayload) => {
+  const { children, items } = payload;
+  let nextByPath = {};
+  Object.values(children).forEach((children) => {
+    Object.assign(
+      nextByPath,
+      createLookupTable(parseSandBoxItemToDetailedItem(children as SandboxItem[], state.itemsByPath), 'path')
+    );
+    if (children.levelDescriptor) {
+      nextByPath[children.levelDescriptor.path] = parseSandBoxItemToDetailedItem(
+        children.levelDescriptor,
+        state.itemsByPath[children.levelDescriptor.path]
+      );
+    }
+  });
+  items.forEach((item) => {
+    nextByPath[item.path] = item;
+  });
+  return { ...state, itemsByPath: { ...state.itemsByPath, ...nextByPath } };
 };
 
 const reducer = createReducer<ContentState>(initialState, (builder) => {
@@ -175,6 +213,7 @@ const reducer = createReducer<ContentState>(initialState, (builder) => {
     }))
     .addCase(pathNavigatorConditionallySetPathComplete, updateItemByPath)
     .addCase(pathNavigatorFetchPathComplete, updateItemByPath)
+    .addCase(pathNavigatorBulkFetchPathComplete, updateItemsByPaths)
     .addCase(pathNavigatorFetchParentItemsComplete, (state, { payload: { items, children } }) => {
       return {
         ...state,
@@ -201,32 +240,28 @@ const reducer = createReducer<ContentState>(initialState, (builder) => {
       };
     })
     .addCase(pathNavigatorTreeFetchPathChildrenComplete, updateItemByPath)
+    .addCase(pathNavigatorTreeBulkFetchPathChildrenComplete, updateItemsByPaths)
     .addCase(pathNavigatorTreeFetchPathPageComplete, updateItemByPath)
     .addCase(
       pathNavigatorTreeRestoreComplete,
       (state, action: { payload: PathNavigatorTreeRestoreCompletePayload }) => {
-        const {
-          payload: { children, items }
-        } = action;
-        let nextByPath = {};
-        Object.values(children).forEach((children) => {
-          Object.assign(
-            nextByPath,
-            createLookupTable(parseSandBoxItemToDetailedItem(children as SandboxItem[], state.itemsByPath), 'path')
-          );
-          if (children.levelDescriptor) {
-            nextByPath[children.levelDescriptor.path] = parseSandBoxItemToDetailedItem(
-              children.levelDescriptor,
-              state.itemsByPath[children.levelDescriptor.path]
-            );
-          }
-        });
-        items.forEach((item) => {
-          nextByPath[item.path] = item;
-        });
-        return { ...state, itemsByPath: { ...state.itemsByPath, ...nextByPath } };
+        const { payload } = action;
+        return updateItemsFromRestoredTree(state, payload);
       }
     )
+    .addCase(pathNavigatorTreeBulkRestoreComplete, (state, { payload: { trees } }) => {
+      let nextByPath = state.itemsByPath;
+      trees.forEach((tree) => {
+        nextByPath = {
+          ...nextByPath,
+          ...updateItemsFromRestoredTree({ ...state, itemsByPath: nextByPath }, tree).itemsByPath
+        };
+      });
+      return {
+        ...state,
+        itemsByPath: nextByPath
+      };
+    })
     .addCase(updateItemsByPath, (state, { payload }) => {
       return updateItemByPath(state, { payload: { parent: null, children: payload.items } });
     })
